config: Show crosstool-NG version in .config
This was previously available in 6c62da480344 but was then removed
in cb7fcbe1ef4f. Add it back since it will be helpful if the user
wants to hide the crosstool-NG version in the pkgversion.

To get the new version, a defconfig from "ct-ng savedefconfig" or
"ct-ng oldconfig" must be used in combination with "ct-ng menuconfig"
or "ct-ng nconfig".

Signed-off-by: Nathan Chancellor <natechancellor@gmail.com>